requirements.txt
langchain
google-generativeai
google-cloud-aiplatform
python-dotenv
streamlit
huggingface_hub